MEMMGR: Fix failure in autoreconf
ACLOCAL_AMFLAGS was probably a left over of a bad copy/paste from a
previous autoconf file. Because the folder 'm4' does not exist for
tiler project, autoreconf will fail to build.

We never noticed the problem so far since we were calling aclocal
directly without using autoreconf, hence this macro was not used!

Signed-off-by: Nicolas Dechesne <n-dechesne@ti.com>